create appointment and medicine app and transfer provider, service, appointment, location model to their apps
