+++ digital_appointment/accounts/models.py
@@ -30,47 +30,3 @@ class User(AbstractBaseUser):
 
     def has_module_perms(self, app_label):
         return True
-
-
-class Appointment(models.Model):
-    STATUS_CHOICES = (
-        ('scheduled', 'scheduled'),
-        ('canceled', 'canceled'),
-        ('completed', 'completed'),
-        ('other', 'other'),
-    )
-
-    user_id = models.ForeignKey(User, on_delete=models.CASCADE)
-    provider_id = models.ForeignKey('Provider', on_delete=models.CASCADE)
-    service_id = models.ForeignKey('Service', on_delete=models.CASCADE)
-    appointment_date = models.DateTimeField()
-    status = models.CharField(max_length=20, choices=STATUS_CHOICES, default='scheduled')
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-
-class Service(models.Model):
-    name = models.CharField(max_length=100)
-    description = models.TextField()
-    duration = models.DurationField()
-    price = models.DecimalField(max_digits=5, decimal_places=2)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-
-class Provider(models.Model):
-    user_id = models.OneToOneField(User, on_delete=models.CASCADE)
-    specialty = models.CharField(max_length=100)
-    location_id = models.ForeignKey('Location', on_delete=models.CASCADE)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
-
-
-class Location(models.Model):
-    name = models.CharField(max_length=100)
-    address = models.CharField(max_length=100)
-    city = models.CharField(max_length=100)
-    state = models.CharField(max_length=100)
-    zip_code = models.CharField(max_length=100)
-    created_at = models.DateTimeField(auto_now_add=True)
-    updated_at = models.DateTimeField(auto_now=True)
