+++ digital_appointment/appointment/apps.py
@@ -1,0 +1,6 @@
+from django.apps import AppConfig
+
+
+class AppointmentConfig(AppConfig):
+    default_auto_field = 'django.db.models.BigAutoField'
+    name = 'appointment'
