+++ digital_appointment/digital_appointment/settings.py
@@ -37,7 +37,9 @@ INSTALLED_APPS = [
     'django.contrib.sessions',
     'django.contrib.messages',
     'django.contrib.staticfiles',
-    'accounts.apps.AccountsConfig'
+    'accounts.apps.AccountsConfig',
+    'medicine.apps.MedicineConfig',
+    'appointment.apps.AppointmentConfig',
 ]
 
 MIDDLEWARE = [
