+++ digital_appointment/medicine/models.py
@@ -1,0 +1,29 @@
+from django.db import models
+from accounts.models import User
+
+
+class Service(models.Model):
+    name = models.CharField(max_length=100)
+    description = models.TextField()
+    duration = models.DurationField()
+    price = models.DecimalField(max_digits=5, decimal_places=2)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+
+class Provider(models.Model):
+    user_id = models.OneToOneField(User, on_delete=models.CASCADE)
+    specialty = models.CharField(max_length=100)
+    location_id = models.ForeignKey('Location', on_delete=models.CASCADE)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
+
+
+class Location(models.Model):
+    name = models.CharField(max_length=100)
+    address = models.CharField(max_length=100)
+    city = models.CharField(max_length=100)
+    state = models.CharField(max_length=100)
+    zip_code = models.CharField(max_length=100)
+    created_at = models.DateTimeField(auto_now_add=True)
+    updated_at = models.DateTimeField(auto_now=True)
